requirements.txt
customtkinter==5.1.2
imageio==2.27.0
ipython==8.12.0
keras==2.12.0
matplotlib==3.6.3
numpy==1.23.5
Pillow==9.5.0
setuptools==65.5.1
tensorflow==2.12.0
tqdm==4.65.0
